requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
tensorflow-macos
tensorflow-metal
keras-tuner
keras